Successfully connected to Google Play Games Services
Game now connects to Google Play Services. The game now runs in a fragment activity instead of an activity as this is required in order to use GoogleApiClient.Builder.enableAutoManage. Should not affect the game in any way. Logics related to Google Play Services should probably be separated into a separate class as more functionality is added.